Initial commit of Jenkinsfile
Adds the ability to manage CI jobs in jenkins via the kubernetes plugin.
The plugin is technically compatible with freestyle jobs, but only with
a limited set of features, so the plan is to switch to using multibranch
pipelines for each project, which will create new jobs for each github
branch, and use a Jenkinsfile for job configuration rather than relying
on lakitu to generate this for each new job/branch.

A few other things to note...

The container resource limits/requests are an initial best guess at
values that will prevent any pods created by jenkins from overloading
our k8s cluster.  The values chosen were based on a mix of testing and
guesswork, and are expected to be updated once we have more data in the
future.

Secrets required for chef cookbook CI builds are 'chef', which includes
the 'jenkins-agent' chef client key required for connecting to the chef
server, and the 'ssh' and 'test-kitchen' secrets, which are likely only
temporary and will be replaced by IAM roles added to jenkins when we
move from using google for test-kitchen to AWS.

[ID-1574]